streamlit_app/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
requests>=2.31.0

